\name{cotton}
\alias{cotton}
\title{Example datasets}
\description{
	\code{cotton} has two objects. \code{cotton_tree} is an object of class \code{"phylo"}, containing a phylogenetic tree for 37 species of cotton (\emph{Gossypium}). \code{cotton_data} is an object of class \code{"matrix"} containing traits for each species. Most trait values are real, but a few values were not available and so random values were used so there wer no NA's in the dataset.
}
\usage{
data(cotton)
cotton_data
cotton_tree
}
\format{
	The data are stored as an object of class \code{"phylo"}, and \code{"matrix"}.
}
\source{
	From the paper: Chamberlain, S.A., & J.A. Rudgers. 2012. How do plants balance multiple mutualists? Correlations among traits for attracting protective bodyguards and pollinators in cotton (Gossypium). Evolutionary Ecology 26:65-77. DOI: 10.1007/s10682-011-9497-3.
}
\keyword{datasets}
