\name{primates}
\alias{primates}
\title{Example DNA dataset from primates}
\description{
	An object of class \code{"DNAbin"} containing nucleotide sequence data of mysterious origin for 12 species of primates.
}
\usage{
data(primates)
}
\format{
	The data are stored as a modified object \code{"DNAbin"}.
}
\source{
	Unknown.
}
\keyword{datasets}
