\name{Rphylip-package}
\alias{Rphylip-package}
\alias{Rphylip}
\docType{package}
\title{
	Rphylip: An R interface for PHYLIP
}
\description{
	\pkg{Rphylip} provides an R interface for programs in the PHYLIP phylogeny methods package (Felsenstein 1989, 2013).
}
\details{
	The complete list of functions can be displayed with \code{library(help = Rphylip)}.

	Obviously, before any of the functions of this package can be used, users must first install PHYLIP (Felsenstein 2013). More information about installing PHYLIP can be found on the PHYLIP webpage: \url{http://evolution.genetics.washington.edu/phylip.html}.

	More information on \pkg{Rphylip} can be found at \url{http://www.phytools.org/Rphylip/} or \url{http://blog.phytools.org}. The latest code for the development version of Rphylip can also be found on github at the following URL: \url{http://github.com/liamrevell/Rphylip}.
}
\author{
	Liam J. Revell, Scott A. Chamberlain

	Maintainer: Liam J. Revell <liam.revell@umb.edu>
}
\references{
	Felsenstein, J. (1989) PHYLIP--Phylogeny Inference Package (Version 3.2). \emph{Cladistics}, 5, 164-166.

	Felsenstein, J. (2013) PHYLIP (Phylogeny Inference Package) version 3.695. Distributed by the author. Department of Genome Sciences, University of Washington, Seattle.

	Revell, L. J. (2013) Rphylip: An R interface for PHYLIP. R package version 0-1.09.
}
\keyword{package}
