\name{as.proseq}
\alias{as.proseq}
\title{Converts objects to protein sequences}
\usage{
as.proseq(x, ...)
}
\arguments{
	\item{x}{an object containing amino sequences. (Presently only objects of class \code{"phyDat"} are permitted.)}
	\item{...}{optional arguments.}
}
\description{
	Converts objects to class \code{"proseq"}.
}
\value{
	An object of class \code{"proseq"} containing protein sequences.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
	\code{\link{print.proseq}}, \code{\link{Rproml}}
}
\keyword{utilities}
\keyword{amino acid}

