\name{primates}
\alias{primates}
\alias{chloroplast}
\alias{primates.bin}
\title{Example datasets}
\description{
	\code{primates} is an object of class \code{"DNAbin"} containing nucleotide sequence data of mysterious origin for 12 species of primates. \code{chloroplast} is an object of class\code{"proseq"} containing a chloroplast alignment from the phangorn package (Schliep 2011). \code{primates.bin} is a an object of class \code{"phylip.data"} containing \code{"0"} and \code{"1"} character data for primates.
}
\usage{
data(primates)
data(chloroplast)
data{primates.bin}
}
\format{
	The data are stored as an object of class \code{"DNAbin"}, \code{"proseq"}, or \code{"phylip.data"}.
}
\source{
	Unknown.
}
\keyword{datasets}
