\name{primates}
\alias{primates}
\alias{chloroplast}
\title{Example datasets}
\description{
	\code{primates} is an object of class \code{"DNAbin"} containing nucleotide sequence data of mysterious origin for 12 species of primates. \code{chloroplast} is an object of class\code{"proseq"} containing a chloroplast alignment from the phangorn package (Schliep 2011).
}
\usage{
data(primates)
data(chloroplast)
}
\format{
	The data are stored as an object of class \code{"DNAbin"} or \code{"proseq"}.
}
\source{
	Unknown.
}
\keyword{datasets}
