\name{print.proseq}
\alias{print.proseq}
\title{Print method protein sequences}
\usage{
\method{print}{proseq}(x, printlen=6, digits=3, ...)
}
\arguments{
	\item{x}{an object of class \code{"proseq"}.}
	\item{printlen}{number of sequence names to print.}
	\item{digits}{number of digits to print.}
	\item{...}{optional arguments.}
}
\description{
	Print method for an object of class \code{"proseq"}.
}
\value{
	Prints to screen.
}
\author{Liam Revell \email{liam.revell@umb.edu}}
\seealso{
	\code{\link{as.proseq}}, \code{\link{Rproml}}
}
\keyword{phylogenetics}
\keyword{inference}
\keyword{maximum likelihood}

